Fix PR issues * Rename `recover_swap.rs` to `recreate_swap_data.rs` * Handle error events
